fix(module9-task1): changes after reviews
