+++ index.html
@@ -23,7 +23,7 @@
 
       <!-- Фильтрация объявлений -->
       <div class="map__filters-container">
-        <form action="https://28.javascript.pages.academy/keksobooking" class="map__filters" method="post" enctype="multipart/form-data" autocomplete="off">
+        <form class="map__filters" autocomplete="off">
           <select name="housing-type" id="housing-type" class="map__filter">
             <option value="any" selected>Любой тип жилья</option>
             <option value="bungalow">Бунгало</option>
