+++ js/const.js
@@ -1,4 +1,5 @@
 const SIMILAR_OFFERS_COUNT = 10;
+const MAX_PRICE_PER_NIGHT = 100000;
 const TITLES = ['Квартира', 'Бунгало', 'Отель'];
 const DESCRIPTIONS = ['Описание помещения 01', 'Описание помещения 02', 'Описание помещения 03'];
 const PHOTOS = [
@@ -16,5 +17,22 @@ const offerTypeToTitle = {
   palace: 'Дворец',
   hotel: 'Отель',
 };
+const Title = {
+  MIN: 30,
+  MAX: 100,
+};
+const roomsCapacity = {
+  '1': ['1'],
+  '2': ['2', '1'],
+  '3': ['3', '2', '1'],
+  '100': ['0'],
+};
+const minPrice = {
+  bungalow: 0,
+  flat: 1000,
+  hotel: 3000,
+  house: 5000,
+  palace: 10000,
+};
 
-export { SIMILAR_OFFERS_COUNT, TITLES, DESCRIPTIONS, PHOTOS, TYPES, TIMES, FEATURES, offerTypeToTitle };
+export { SIMILAR_OFFERS_COUNT, TITLES, DESCRIPTIONS, PHOTOS, TYPES, TIMES, FEATURES, offerTypeToTitle, Title, MAX_PRICE_PER_NIGHT, roomsCapacity, minPrice };
