+++ js/filter.js
@@ -5,12 +5,12 @@ const mapFiltersElements = mapFilters.querySelectorAll('select, fieldset');
 
 const setFiltersActive = () => {
   mapFilters.classList.add('map__filters--disabled');
-  setElementsDisabled(mapFiltersElements, true);
+  setElementsDisabled(mapFiltersElements, false);
 };
 
 const setFiltersInactive = () => {
   mapFilters.classList.remove('map__filters--disabled');
-  setElementsDisabled(mapFiltersElements, false);
+  setElementsDisabled(mapFiltersElements, true);
 };
 
 export { setFiltersActive, setFiltersInactive };
